+++ src/model/Device.h
@@ -19,7 +19,7 @@ public:
     long getForcedUntil();
     unsigned short getBrightnessMax();
     unsigned short getBrightnessMin();
-    unsigned short getHeatingAnticipation();
+    float getHeatingAnticipation();
 
     void setPowerOn(bool powerOn);
     void setForced(bool forced);
@@ -27,7 +27,7 @@ public:
     void setForcedUntil(long forcedUntil);
     void setBrightnessMax(unsigned short brightnessMax);
     void setBrightnessMin(unsigned short brightnessMin);
-    void setHeatingAnticipation(unsigned short heatingAnticipation);
+    void setHeatingAnticipation(float heatingAnticipation);
 
 private:
     bool _powerOn;
@@ -36,7 +36,7 @@ private:
     long _forcedUntil;
     unsigned short _brightnessMax;
     unsigned short _brightnessMin;
-    unsigned short _heatingAnticipation;
+    float _heatingAnticipation;
 };
 
 #endif
